+++ src/app/(afterLogin)/(main)/home/daily/[day]/_components/daily/DailyNoThread.tsx
