+++ src/app/(afterLogin)/(main)/home/daily/[day]/_components/daily/DailyThread.tsx
