+++ src/app/_components/common/bottomMenuBar/BottomMenuBar.tsx
@@ -15,7 +15,7 @@ export default function BottomMenuBar() {
   return (
     <nav className="sticky bottom-0">
       <div className="flex justify-evenly items-start pt-[1.3rem] bg-white w-[37.5rem] h-[4.8rem] border-bgGrey border-t">
-        <Link href={`/daily/${date}`}>
+        <Link href={`/home/daily/${date}`}>
           <HomeIcon isActive={pathName.includes('daily')} />
         </Link>
         <Link href="/favorites">
